// src/types.ts
export interface Feature {
  titile: string;
  subTitle: string;
  image: any;
}
export interface VerticalImageSliderProps {
  image: any;
}
export interface whyInvest {
  title: string;
  descripetion: string;
}
export interface teamType {
  name: string;
  role: string;
  image: any;
  social: {
    linkedin: string;
  };
}
export interface ComparisonType {
  name: string;
  icon: any;
  defytrends: boolean;
  dash2trade: boolean;
  lunarcrush: boolean;
}